requirements-dev.txt
mr.bob==0.1a9
PyYAML==3.12
click==6.6
GitPython==1.0.1
Sphinx==1.3.3
alabaster==0.7.6
tox==2.3.1
nose==1.3.7
watchdog==0.8.3
